Split scheduler views file by class, reformat the split files
