+++ csm_web/scheduler/views/__init__.py
@@ -1,0 +1,7 @@
+from .course import CourseViewSet
+from .profile import ProfileViewSet
+from .resource import ResourceViewSet
+from .section import SectionViewSet
+from .spacetime import SpacetimeViewSet
+from .student import StudentViewSet
+from .user import UserViewSet
