+++ csm_web/scheduler/views/user.py
@@ -1,0 +1,20 @@
+from rest_framework.exceptions import PermissionDenied
+from rest_framework.response import Response
+
+from .utils import viewset_with
+from ..models import Coordinator, User
+
+
+class UserViewSet(*viewset_with("list")):
+    serializer_class = None
+    queryset = User.objects.all()
+
+    def list(self, request):
+        if not (
+            request.user.is_superuser
+            or Coordinator.objects.filter(user=request.user).exists()
+        ):
+            raise PermissionDenied(
+                "Only coordinators and superusers may view the user email list"
+            )
+        return Response(self.queryset.order_by("email").values_list("email", flat=True))
